+++ src/main/java/lotto/model/calculator/LottoRateCalculator.java
@@ -1,4 +1,4 @@
-package lotto.calculator;
+package lotto.model.calculator;
 
 import java.util.List;
