+++ src/main/java/lotto/model/lottos/LottoNumbersGenerator.java
@@ -1,4 +1,4 @@
-package lotto.lottos;
+package lotto.model.lottos;
 
 import camp.nextstep.edu.missionutils.Randoms;
