+++ src/main/java/lotto/model/lottos/Lottos.java
@@ -1,4 +1,4 @@
-package lotto.lottos;
+package lotto.model.lottos;
 
 import java.util.ArrayList;
 import java.util.List;
